requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et-xmlfile==2.0.0
fastapi==0.115.12
jinja2==3.1.6
itsdangerous==2.2.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
playwright==1.49.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-telegram-bot==21.6
pytz==2025.2
requests==2.32.4
schedule==1.2.2
six==1.17.0
slack_sdk==3.35.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
pytest==8.0.0
